Rename concat as followedBy
Matches the naming of `Iterable.followedBy`. Leave concat as a
deprecated alias.
