+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: stream_transform
 description: A collection of utilities to transform and manipulate streams.
 author: Dart Team <misc@dartlang.org>
 homepage: https://www.github.com/dart-lang/stream_transform
-version: 0.0.10
+version: 0.0.11
 
 environment:
   sdk: ">=2.0.0-dev.20.0 <2.0.0"
